Introduction of tleforger and antenna modules
This commit introduces two new modules for the scepter: the tleforger and the antenna ones.
The tleforger module allows to create artificial TLEs that could be used, for example, for sharing and compatibility studies in the ITU-R.
The antenna module is planned to be a collection of various antenna modules. The first one to be introduced is an implementation of ITU-R Recommendation S.1528-0 recommends 1.2.
These modules were designed to be able to integrate with astropy, cysgp4 and pycraf.