Updating skynet module and introducing the earthgrid module (#3)
The proposed changes include an update to the Skynet module to have a new function called pointgen_S_1586_1. The main difference with the existing pointgen is strict compliance with the sky cell division example in ITU-R Rec. S.1586-1. It was found that pointgen creates 2292 sky cells, while Rec. Table 1 has 2334 sky cells. While pointgen implementation seems to be implementing the Rec. methodology more accurately, Table 1 is the de facto standard in spectrum management.
The Earthgrid module provides functions to calculate satellite footprint as well as to divide the Earth into hexagonal cells. Cell centres are returned as they might be used for beam pointing in simulations. Also, a function to reduce cells to only impactful is also introduced.